updated for no network dialog and removed splash references.
