+++ app/src/main/java/com/trailbook/kole/activities/TrailBookActivity.java
@@ -17,8 +17,6 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.res.Resources;
 import android.location.Location;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
@@ -114,6 +112,8 @@ public class TrailBookActivity extends Activity
     public static final String PATH_DETAILS_DIALOG_TAG = "path_details_dialog";
     public static final String FILTER_DIALOG_TAG = "FILTER_DIALOG";
     public static final String INITIAL_PATH_ID_KEY = "INITIAL_PATH_ID";
+    private static final long SPLASH_DISPLAY_LENGTH = 2000;
+    public static final String TAG_SPLASH_SCREEN = "SPLASH_FRAG";
 
     private CharSequence mTitle; // Used to store the last screen title. For use in {@link #restoreActionBar()}.
     private TrailBookMapFragment mMapFragment;
@@ -135,6 +135,7 @@ public class TrailBookActivity extends Activity
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
         requestWindowFeature(Window.FEATURE_PROGRESS);
 
@@ -344,12 +345,12 @@ public class TrailBookActivity extends Activity
         } else if (action == ApplicationUtils.MENU_CONTEXT_DOWNLOAD_ID) {
             onDownloadRequested(pathId);
         } else if (action == ApplicationUtils.MENU_CONTEXT_DELETE_FROM_CLOUD_ID) {
-            if (isNetworkConnected()) {
+            if (ApplicationUtils.isNetworkConnected(this)) {
                 Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " Confirming delete path:" + pathId);
                 deleteFromCloud(pathId);
             } else {
                 Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " no network connectivity");
-                showNoNetworkStatusDialog();
+                ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_message));
             }
         } else if (action == ApplicationUtils.MENU_CONTEXT_RESUME_ID) {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + ": resuming path " + pathId);
@@ -766,13 +767,13 @@ public class TrailBookActivity extends Activity
 
     @Override
     public void onDownloadRequested(String pathId) {
-        if (isNetworkConnected()) {
+        if (ApplicationUtils.isNetworkConnected(this)) {
             mWorkFragment.startDownloadPath(pathId);
             mMapFragment.displayMessage(getString(R.string.download_in_progress));
             collapseSlidingPanel();
         }else {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " no network connectivity");
-            showNoNetworkStatusDialog();
+            ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_message));
         }
     }
 
@@ -988,7 +989,7 @@ public class TrailBookActivity extends Activity
     }
 
     private void refreshFromCloudIfConnectedToNetwork() {
-        if (isNetworkConnected()) {
+        if (ApplicationUtils.isNetworkConnected(this)) {
             Log.d(Constants.TRAILBOOK_TAG, getLocalClassName() + ": refreshing from cloud");
             mWorkFragment.startGetPathSummariesRemote(null, 0);
         } else {
@@ -1041,27 +1042,6 @@ public class TrailBookActivity extends Activity
         ApplicationUtils.showAlert(this, clickListenerOK, getString(R.string.confirm_exit_title), reason, getString(R.string.exit_confirm), getString(R.string.cn_cancel));
     }
 
-    private boolean isNetworkConnected() {
-        ConnectivityManager connMgr = (ConnectivityManager)
-                getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo networkInfo = connMgr.getActiveNetworkInfo();
-        if (networkInfo != null && networkInfo.isConnected()) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    public void showNoNetworkStatusDialog() {
-        DialogInterface.OnClickListener clickListenerOK = new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog,int id) {
-                Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + ": dismissing no paths dialog");
-            }
-        };
-        ApplicationUtils.showAlert(this, clickListenerOK, getString(R.string.no_network_title), getString(R.string.no_network_message), getString(R.string.OK), null);
-    }
-
-
     private void showErrorParsingKMLAlert() {
         DialogInterface.OnClickListener clickListenerOK = new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog,int id) {
@@ -1095,6 +1075,7 @@ public class TrailBookActivity extends Activity
 
     private void toastNoNetwork() {
         Toast.makeText(this, getString(R.string.no_network_toast), Toast.LENGTH_LONG).show();
+        ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_at_startup));
     }
 
     @Override
@@ -1119,7 +1100,7 @@ public class TrailBookActivity extends Activity
     }
 
     public void upload(String pathId) {
-        if (isNetworkConnected()) {
+        if (ApplicationUtils.isNetworkConnected(this)) {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " Uploading path:" + pathId);
             //TODO: update path details and confirm
             Action uploadAction = new PathUploaderAction(mWorkFragment, mPathManager.getPathSummary(pathId));
@@ -1134,7 +1115,7 @@ public class TrailBookActivity extends Activity
             }
         } else {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " no network connectivity");
-            showNoNetworkStatusDialog();
+            ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_message));
         }
     }
 
@@ -1170,7 +1151,7 @@ public class TrailBookActivity extends Activity
         mPathManager.saveComment(comment);
         BusProvider.getInstance().post(new PathCommentAddedEvent(comment));
         if (mPathManager.isPathInCloudCache(comment.getPathId())) {
-            if (isNetworkConnected()) {
+            if (ApplicationUtils.isNetworkConnected(this)) {
                 Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " Uploading comment:" + comment);
                 //TODO: show progress dialog
                 Action uploadAction = new CommentUploaderAction(mWorkFragment, comment);
@@ -1187,14 +1168,14 @@ public class TrailBookActivity extends Activity
             } else {
                 //todo: create service to upload when back on line
                 Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " no network connectivity");
-                showNoNetworkStatusDialog();
+                ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_message));
             }
         }
     }
 
     @Override
     public void onNewAttachedCommentClick(TrailBookComment comment) {
-        if (isNetworkConnected()) {
+        if (ApplicationUtils.isNetworkConnected(this)) {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " Uploading comment:" + comment);
             //TODO: show progress dialog
             PointAttachedObject paoComment = attachNoteToCurrentLocation(comment);
@@ -1210,7 +1191,7 @@ public class TrailBookActivity extends Activity
             }
         } else {
             Log.d(Constants.TRAILBOOK_TAG, CLASS_NAME + " no network connectivity");
-            showNoNetworkStatusDialog();
+            ApplicationUtils.showNoNetworkStatusDialog(this, getString(R.string.no_network_message));
         }
     }
